refactor: create LinkTokenManager base class for managers that require a link token
