+++ fintoc/managers/accounts_manager.py
@@ -1,18 +1,14 @@
 """Module to hold the accounts manager."""
 
-from fintoc.mixins import ManagerMixin
+from fintoc.managers._link_token_manager import LinkTokenManager
 
 
-class AccountsManager(ManagerMixin):
+class AccountsManager(LinkTokenManager):
 
     """Represents an accounts manager."""
 
     resource = "account"
     methods = ["all", "get"]
-
-    def __init__(self, path, client, link_token=None):
-        super().__init__(path, client)
-        self._link_token = link_token
 
     def post_get_handler(self, object_, identifier, **kwargs):
         # pylint: disable=protected-access
@@ -32,9 +28,3 @@ class AccountsManager(ManagerMixin):
                     yield obj
 
             return modified_generator()
-
-    def _all(self, **kwargs):
-        return super()._all(**{**kwargs, "link_token": self._link_token})
-
-    def _get(self, identifier, **kwargs):
-        return super()._get(identifier, **{**kwargs, "link_token": self._link_token})
